arm: dts: am62a: allow booting from eMMC
The bootph-all property in u-boot enables driver initialization prior to
relocation, this is necessary to use the device as boot media.

sdhci0 is the phandle for eMMC on am62a, so this change allows us to use
eMMC as a boot media.

Signed-off-by: Anshul Dalal <anshuld@ti.com>
(cherry picked from commit 243a6dbfb680f9773e3c2a1e2e17e69068830666)