configs: mvebu_espressobin_ultra-88f3720_defconfig: enable hwrng
This device has a hardware random number generator. Linux can
use this feature to randomize the location of the kernel in
memory for better security. However, that functionality is only
available if the bootloader firmware provides it. Enable support
for it in the default configuration for this device.

Signed-off-by: Benjamin Schneider <ben@bens.haus>
Reviewed-by: Stefan Roese <sr@denx.de>
(cherry picked from commit e923dffb63d8600b345eaa38737bee7fbc5aa649)